backend/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
google-adk>=0.1.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0 